[refactor] seat-hold·seat-selection 도메인 entities/features 승격 #114
