+++ src/app/providers/router/ui/AppRouter.tsx
@@ -28,12 +28,12 @@ import PaymentProcessingPage from '@/pages/tickets/ui/payment/PaymentProcessingP
 import PaymentCompletePage from '@/pages/tickets/ui/payment/PaymentCompletePage';
 import BooksPage from '@/pages/books';
 import SeatsPage from '@/pages/books/ui/SeatsPage';
+import SeatHoldLifecycleController from '@/features/seat-booking/ui/SeatHoldLifecycleController';
 import { MypagePage, AccountPage, PurchaseDetailPage, SaleDetailPage } from '@/pages/mypage';
 import MypageLayout from '@/shared/widgets/layout/mypage/MypageLayout';
 import AuthSessionController from './AuthSessionController';
 import BookingFlowStateGuard from './BookingFlowStateGuard';
 import OAuthMessageListener from './OAuthMessageListener';
-import SeatHoldLifecycleController from './SeatHoldLifecycleController';
 
 const AppRouter = () => {
    return (
