+++ src/entities/seat-selection/model/constants.ts
@@ -1,0 +1,1 @@
+export const MAX_SELECTED_SEATS = 4;
