+++ src/pages/books/ui/SeatsPage.tsx
@@ -6,9 +6,9 @@ import { createSeatsForZone } from '@/pages/books/model/seatData';
 import { getResellZoneInsights, type ResellListingItem } from '@/pages/books/model/resellData';
 import { getSelectedSeatDetails } from '@/pages/books/model/selectedSeats';
 import { useSeatHoldActions } from '@/pages/books/model/useSeatHoldActions';
-import { useSeatHoldStore } from '@/pages/books/model/useSeatHoldStore';
+import { useSeatHoldStore } from '@/entities/seat-hold/model/useSeatHoldStore';
+import { useSeatSelectionStore } from '@/entities/seat-selection/model/useSeatSelectionStore';
 import { useSeatMapData } from '@/pages/books/model/useSeatMapData';
-import { useSeatSelectionStore } from '@/pages/books/model/useSeatSelectionStore';
 import { formatPrice, getBookingZones, getZoneOverviewImage, getStadiumName } from '@/pages/books/model/zoneData';
 import type { SeatItem } from '@/pages/books/model/types';
 import { getBookingFlowMode } from '@/shared/lib/booking-flow';
