+++ src/pages/tickets/ui/payment/TicketPaymentPage.tsx
@@ -4,8 +4,8 @@ import { useEffect, useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { getSelectedSeatPaymentSummary } from '@/pages/books/model/getSelectedSeatPaymentSummary';
 import { getSelectedSeatDetails } from '@/pages/books/model/selectedSeats';
-import { useSeatHoldStore } from '@/pages/books/model/useSeatHoldStore';
-import { useSeatSelectionStore } from '@/pages/books/model/useSeatSelectionStore';
+import { useSeatHoldStore } from '@/entities/seat-hold/model/useSeatHoldStore';
+import { useSeatSelectionStore } from '@/entities/seat-selection/model/useSeatSelectionStore';
 import { getBookingTeamConfig, getBookingZones } from '@/pages/books/model/zoneData';
 import { Button } from '@/shared/ui/button';
 import { useBookingEntryStore, type BookingEntryState } from '@/shared/lib/useBookingEntryStore';
